Add xquad task
