+++ lm_eval/tasks/spanish_bench/utils.py
@@ -35,15 +35,6 @@ def process_doc_nli(dataset):
     return dataset.map(process_fn)
 
 
-def process_results_qa(doc, results):
-    preds = results[0]
-    reference = doc["answers"]["text"][0]
-    # import code; code.interact(local=dict(globals(), **locals()))
-    f1_sum = squad_metrics.compute_f1(reference, preds)
-    exact_match = squad_metrics.compute_exact(reference, preds)
-    return {"f1": f1_sum, "exact_match": exact_match}
-
-
 def process_xlsum(dataset):
     def _process_doc(doc):
         # Remove double spaces
