+++ lm_eval/tasks/xquad/xquad_ar.yaml
@@ -1,0 +1,4 @@
+include: xquad_common_yaml
+task: xquad_ar
+dataset_name: xquad.ar
+doc_to_text: "سيا: {{context}}\n\nسؤال: {{question}}\n\nإجابة:"
