+++ lm_eval/tasks/xquad/xquad_common_yaml
@@ -1,12 +1,16 @@
-task: xquad_es
+# This file will be included in the generated language-specific task configs.
+# It doesn't have a yaml file extension as it is not meant to be imported directly
+# by the harness.
+tag: xquad
+task: null
 dataset_path: xquad
-dataset_name: xquad.es
+dataset_name: null
 output_type: generate_until
-doc_to_text: "Contexto: {{context}}\n\nPregunta: {{question}}\n\nRespuesta:"
-doc_to_target: '{{answers["text"][0]}}'
 validation_split: validation
-target_delimiter: ' '
+doc_to_text: null
+doc_to_target: '{{answers["text"][0]}}'
 process_results: !function utils.process_results_qa
+target_delimiter: ' '
 generation_kwargs:
   until:
     - "\n"
