+++ lm_eval/tasks/xquad/xquad_hi.yaml
@@ -1,0 +1,4 @@
+include: xquad_common_yaml
+task: xquad_hi
+dataset_name: xquad.hi
+doc_to_text: "प्रसंग: {{context}}\n\nसवाल: {{question}}\n\nउत्तर:"
