+++ lm_eval/tasks/xquad/xquad_vi.yaml
@@ -1,0 +1,4 @@
+include: xquad_common_yaml
+task: xquad_vi
+dataset_name: xquad.vi
+doc_to_text: "Bối cảnh: {{context}}\n\nCâu hỏi: {{question}}\n\nTrả lời:"
